+++ cai_causal_graph/time_series_causal_graph.py
@@ -120,17 +120,16 @@ class TimeSeriesCausalGraph(CausalGraph):
         self._minimal_graph: Optional[TimeSeriesCausalGraph] = None
 
     # Overwrite type annotations for linting and code completion.
-    from_adjacency_matrix: Callable[
+    from_adjacency_matrix: Callable[  # type: ignore
         [
-            Arg(Type[CausalGraph], 'cls'),
             Arg(numpy.ndarray, 'adjacency'),
             DefaultArg(Optional[List[Union[NodeLike, int]]], 'node_names'),
         ],
         TimeSeriesCausalGraph,
     ]
-    from_skeleton: Callable[[Arg(Type[CausalGraph], 'cls'), Arg(Skeleton, 'skeleton')], TimeSeriesCausalGraph]
-    from_networkx: Callable[[Arg(Type[CausalGraph], 'cls'), Arg(networkx.Graph, 'g')], TimeSeriesCausalGraph]
-    from_gml_string: Callable[[Arg(Type[CausalGraph], 'cls'), Arg(str, 'gml')], TimeSeriesCausalGraph]
+    from_skeleton: Callable[[Arg(Skeleton, 'skeleton')], TimeSeriesCausalGraph]  # type: ignore
+    from_networkx: Callable[[Arg(networkx.Graph, 'g')], TimeSeriesCausalGraph]  # type: ignore
+    from_gml_string: Callable[[Arg(str, 'gml')], TimeSeriesCausalGraph]  # type: ignore
 
     def __eq__(self, other: object, deep: bool = False) -> bool:
         """
@@ -1069,6 +1068,11 @@ class TimeSeriesCausalGraph(CausalGraph):
         ), f'The shape of all the adjacency matrices must be the same. Got the following shapes: {list(set(shapes))}.'
         shape = shapes[0]
 
+        # if 0 in adjacency_matrices keys, then we need to add the contemporaneous nodes
+        if 0 not in adjacency_matrices:
+            adjacency_matrices = adjacency_matrices.copy()
+            adjacency_matrices[0] = numpy.zeros((shape[0], shape[0]))
+
         if variable_names is not None:
             variable_names_str: List[Union[str, int]] = []
             assert len(variable_names) == shape[0], (
@@ -1089,28 +1093,29 @@ class TimeSeriesCausalGraph(CausalGraph):
         # delta but if we have many time deltas, this could be very memory intensive. Therefore, we create the graph by
         # adding the edges one by one for each time delta.
 
-        # create the empty graph
-        tsgraph = cls()
+        # create the full matrix from the dictionary of adjacency matrices
+        # get the nodes first the full adjacency matrix has the shape NT x NT, where N is the number of nodes and T
+        # is the number of time steps.
+        adjacency_matrix_full = numpy.zeros((shape[0] * len(adjacency_matrices), shape[0] * len(adjacency_matrices)))
 
-        # first add all the contemporaneous nodes (there could be floating nodes)
+        # create the node names list to match the variable names
+        node_names = []
         for variable_name in variable_names_str:
-            tsgraph.add_node(variable_name=variable_name, time_lag=0)
+            for time_delta in adjacency_matrices:
+                node_names.append(get_name_with_lag(str(variable_name), time_delta))
+
+        # create a map between time delta and the index in the full adjacency matrix
+        time_delta_to_index = {time_delta: i for i, time_delta in enumerate(adjacency_matrices)}
 
         for time_delta, adjacency_matrix in adjacency_matrices.items():
-            # create the edges
-            edges: List[Tuple[str, str]] = []
-            # get the edges from the adjacency matrix by getting the indices of the non-zero elements
             for row, column in zip(*numpy.where(adjacency_matrix)):
-                edges.append(
-                    (
-                        get_name_with_lag(variable_names_str[row], time_delta),
-                        variable_names_str[column],
-                    )
-                )
-            # add the edges to the graph
-            tsgraph.add_edges_from(edges)  # type: ignore
+                # add 1 to the row and column to account for the time delta
+                adjacency_matrix_full[
+                    time_delta_to_index[time_delta] + ((shape[0] - 1) * row),
+                    time_delta_to_index[0] + ((shape[0] - 1) * column),
+                ] = 1
 
-        return tsgraph
+        return cls.from_adjacency_matrix(adjacency_matrix_full, node_names)  # type: ignore
 
     @property
     def adjacency_matrices(self) -> Dict[int, numpy.ndarray]: