+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## NEXT
+
+- Fixed a bug in `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.from_adjacency_matrices` for
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` where the method was not handling properly
+  undirected edges in the adjacency matrices.
+- Added `mypy-extensions` dependency as `"^1.0.0"` and moved `mypy` back to dev dependency (setting it as `"^1.8.0"`).
+
 ## 0.4.4
 
 - Relaxed `mypy` dependency to `"*"`.